Move desktop update types out of the upstream api-types module
lib/api-types.ts is owned by agegr/pi-web and holds its skill and plugin API
types. The fork appended five update-check types to it, which put a file
upstream extends on its own schedule permanently on the drifted list — for no
benefit, since every consumer of those types (the updates route, app-updates,
UpdateReminder, AppSettings) is fork-owned.

They now live in lib/app-update-types.ts and lib/api-types.ts is byte-identical
to upstream again, so it drops off the boundary manifest entirely. One fewer
file that can block an unattended sync, and one fewer place to resolve a
conflict.

Verified beyond types: /api/updates still returns all three components with
their release status, and the settings panel renders them.
